Fix64ToDouble(), Fix32ToDouble() were incorrectly returning floats. Added unit tests to detect this in future.

+++ src/include/FixedPoint-Fp32f.hpp
@@ -206,7 +206,7 @@ namespace Fp
 	//! @details	Good for debugging fixed-point arithmetic.
 	//! @warning 	Slow!
 	template <uint8_t p>
-	float Fix32ToDouble(int32_t f)
+	double Fix32ToDouble(int32_t f)
 	{
 		return (double)f / (double)(1 << p);
 	}
